+++ Detector/WireCell.hh
@@ -1,7 +1,7 @@
 #ifndef KinKal_WireCell_hh
 #define KinKal_WireCell_hh
 //
-// class describing a wire cell.  Eventually this could have more structure
+// base class describing a wire cell.  Eventually this could have more structure
 //
 #include "KinKal/General/Vectors.hh"
 namespace KinKal {
@@ -19,28 +19,5 @@ namespace KinKal {
       virtual double maximumDriftTime() const = 0; // maximum drift
       // add functions to describe material TODO
   };
-
-  // simple implementation; a realistic implementation needs EXB effects, non-constant drift, etc. TODO
-  class SimpleCell : public WireCell {
-    public:
-      double size() const override { return 2*rcell_; }
-      void distanceToTime(POL2 const& drift, double& tdrift, double& tdriftvar, double& dspeed) const override {
-      tdrift  = drift.R()/dvel_;
-      tdriftvar = tvar_; 
-      dspeed = dvel_;
-    }
-    // provide seed (mm/ns) and time RMS (ns) on construction
-    SimpleCell(double driftspeed, double tvar, double rcell) :dvel_(driftspeed), tvar_(tvar), rcell_(rcell) {}
-    double averageDriftSpeed() const override { return dvel_; }
-    double maximumDriftTime() const override { return rcell_/dvel_; }
-    virtual ~SimpleCell(){}
-    double timeVariance() const { return tvar_; }
-    double radius() const { return rcell_; }
-    private:
-    double dvel_; // constant drift speed
-    double tvar_; // constant time variance
-    double rcell_; // straw radius
-  };
-
 }
 #endif
