+++ Trajectory/LoopHelix.cc
@@ -94,7 +94,7 @@ namespace KinKal {
     LoopHelix(pstate.position4(),pstate.momentum4(),charge,bnom,range) 
   {}
 
-  LoopHelix::LoopHelix(ParticleStateMeasurement const& pstate, int charge, VEC3 const& bnom, TimeRange const& range) :
+  LoopHelix::LoopHelix(ParticleStateEstimate const& pstate, int charge, VEC3 const& bnom, TimeRange const& range) :
   LoopHelix(pstate.stateVector(),charge,bnom,range) {
   // derive the parameter space covariance from the global state space covariance
     PSMAT dpds = dPardState(pstate.stateVector().time());
@@ -349,10 +349,10 @@ namespace KinKal {
     return dsdp;
   }
 
-  ParticleStateMeasurement LoopHelix::measurementState(double time) const {
+  ParticleStateEstimate LoopHelix::stateEstimate(double time) const {
   // express the parameter space covariance in global state space
     PSMAT dsdp = dStatedPar(time);
-    return ParticleStateMeasurement(state(time),ROOT::Math::Similarity(dsdp,pars_.covariance()));
+    return ParticleStateEstimate(state(time),ROOT::Math::Similarity(dsdp,pars_.covariance()));
   }
 
   void LoopHelix::print(ostream& ost, int detail) const {
